Remove unnecessary Reconciler interface
